vit/requirements_v3.txt
absl-py==2.3.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
# detectron2==0.6
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
fvcore==0.1.5.post20221221
grpcio==1.72.1
huggingface-hub==0.32.4
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iopath==0.1.9
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.7
lightning-utilities==0.14.3
markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==1.3.5
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
portalocker==3.1.1
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure-eval==0.2.3
pycocotools==2.0.10
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-ranger==0.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
safetensors==0.5.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
timm==1.0.15
tomli==2.2.1
# torch==2.0.0
torch-optimizer==0.3.0
torchmetrics==1.3.2
# torchvision==0.15.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
werkzeug==3.1.3
yacs==0.1.8
zipp==3.22.0